combined login with journal
